Removed an unnecessary link in afus_owner table. Refactored a common TransformationLog file to encapsulate the errors and warnings for all surveys and updated imports.
