+++ dap-etl/transformation/src/main/scala/org/broadinstitute/monster/dap/common/TransformationLog.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.monster.dap.hles
+package org.broadinstitute.monster.dap.common
 
 import com.spotify.scio.ScioMetrics.counter
 import org.broadinstitute.monster.dap.common.PostProcess.errCount
@@ -9,13 +9,13 @@ import ujson.Obj
 //   To note a skippable error you encountered, you should just call the `log` method of these classes.
 //   They should only be thrown in cases where a missing field means the entire record must be thrown out, and you should
 //     still call `log` in the block where you catch the error and skip the record.
-abstract class HLESurveyTransformationLog extends Exception {
+abstract class TransformationLog extends Exception {
   val jsonMsg: Obj
 
   def log(implicit logger: Logger): Unit
 }
 
-class HLESurveyTransformationError(msg: String) extends HLESurveyTransformationLog {
+class TransformationError(msg: String) extends TransformationLog {
 
   def log(implicit logger: Logger): Unit = {
     logger.error(jsonMsg.toString())
@@ -29,11 +29,11 @@ class HLESurveyTransformationError(msg: String) extends HLESurveyTransformationL
     )
 }
 
-class HLESurveyTransformationWarning(msg: String) extends HLESurveyTransformationLog {
+class TransformationWarning(msg: String) extends TransformationLog {
 
   def log(implicit logger: Logger): Unit = {
     logger.error(jsonMsg.toString())
-    counter("hles", "hles_transformation_warning").inc()
+    counter("DAP Survey", "transformation_warning").inc()
   }
 
   val jsonMsg: Obj = ujson
@@ -44,7 +44,7 @@ class HLESurveyTransformationWarning(msg: String) extends HLESurveyTransformatio
 }
 
 // case classes for all the different actual warnings and errors we want to raise during the workflow
-case class MissingOwnerIdError(msg: String) extends HLESurveyTransformationError(msg)
-case class TruncatedDecimalError(msg: String) extends HLESurveyTransformationWarning(msg)
-case class MissingCalcFieldError(msg: String) extends HLESurveyTransformationError(msg)
-case class InvalidArmMonthError(msg: String) extends HLESurveyTransformationError(msg)
+case class MissingOwnerIdError(msg: String) extends TransformationError(msg)
+case class TruncatedDecimalError(msg: String) extends TransformationWarning(msg)
+case class MissingCalcFieldError(msg: String) extends TransformationError(msg)
+case class InvalidArmMonthError(msg: String) extends TransformationError(msg)
