+++ dap-etl/transformation/src/main/scala/org/broadinstitute/monster/dap/dog/DemographicsTransformations.scala
@@ -1,8 +1,7 @@
 package org.broadinstitute.monster.dap.dog
 
 import java.time.{LocalDate, Period}
-import org.broadinstitute.monster.dap.common.RawRecord
-import org.broadinstitute.monster.dap.hles.MissingCalcFieldError
+import org.broadinstitute.monster.dap.common.{MissingCalcFieldError, RawRecord}
 import org.broadinstitute.monster.dogaging.jadeschema.fragment.HlesDogDemographics
 
 object DemographicsTransformations {
